+++ src/components/MyKiva/LendingStats.vue
@@ -1,13 +1,13 @@
 <template>
 	<div
 		class="tw-mb-2"
-		:class="{'next-steps-link': nextStepsExperimentVariant}"
+		:class="{'next-steps-link': isNextStepsExperimentEnabled}"
 	>
 		<h3 class="tw-text-primary md:tw-mb-1">
 			Next steps recommended for you
 		</h3>
 		<div
-			v-if="nextStepsExperimentVariant"
+			v-if="isNextStepsExperimentEnabled"
 			class="tw-flex md:tw-gap-1 tw-cursor-pointer tw-w-16 md:tw-w-fit tw-justify-end"
 			@click="$router.push('/mykiva/next-steps')"
 		>
@@ -47,7 +47,8 @@
 					:post-lending-next-steps-enable="postLendingNextStepsEnable"
 					:user-info="userInfo"
 					:show-post-lending-next-steps-cards="showPostLendingNextStepsCards"
-					@open-goal-modal="showGoalModal = true"
+					:goal-editing-enable="goalEditingEnable"
+					@open-goal-modal="openGoalModal($event)"
 					@open-impact-insight-modal="showImpactInsightsModal = true"
 				/>
 			</div>
@@ -160,7 +161,7 @@
 		<JourneyCardCarousel
 			v-else
 			class="carousel tw--mt-6"
-			:class="{'carousel-spacing': nextStepsExperimentVariant}"
+			:class="{'carousel-spacing': isNextStepsExperimentEnabled}"
 			user-in-homepage
 			in-lending-stats
 			controls-top-right
@@ -181,6 +182,7 @@
 			:latest-loan="latestLoan"
 			:user-info="userInfo"
 			:show-post-lending-next-steps-cards="showPostLendingNextStepsCards"
+			:goal-editing-enable="goalEditingEnable"
 			@open-goal-modal="openGoalModal($event)"
 			@open-impact-insight-modal="showImpactInsightsModal = true"
 		/>
@@ -193,6 +195,7 @@
 			:is-goal-set="isGoalSet"
 			:show-goal-selector="true"
 			:tiered-achievements="heroTieredAchievements"
+			:is-updating-goal="isUpdatingGoal"
 			@close-goal-modal="closeGoalModal"
 			@set-goal="setGoal"
 		/>
@@ -304,6 +307,10 @@ export default {
 			type: String,
 			default: 'a',
 			validator: value => ['a', 'b'].includes(value)
+		},
+		goalEditingEnable: {
+			type: Boolean,
+			default: false
 		},
 	},
 	data() {